+++ CHANGELOG.md
@@ -4,6 +4,21 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+## [2.46.0] 2021-08-16
+
+- Allow to run git delta command on local updates
+- Update labels of hardis:data commands
+- New technical command: hardis:work:ws , to call VsCode Extension refresh from CLI
+
+## [2.45.0] 2021-08-15
+
+- Refactor **hardis:org:diagnose:legacyapi** with jsforce to handle more log entries
+- Do not display `git branch -v` in logs
+
+## [2.44.0] 2021-08-14
+
+- New command **hardis:org:diagnose:legacyapi** : Detect [use of deprecated API versions](https://help.salesforce.com/articleView?id=000351312&type=1&mode=1&language=en_US) in production org
+
 ## [2.43.1] 2021-07-23
 
 - Update deployTips